requirements.txt
streamlit
opencv-python
openai
numpy
requests
pydub
ffmpeg-python
